@@ -1,6 +1,6 @@
 module tasker
 
-go 1.19
+go 1.20
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.1
